+++ src/classes/web/WsConnect.ts
@@ -11,7 +11,7 @@ export default class WSConnect<Context extends Record<any, any> = {}, Type = 'co
 	/**
 	 * Initializes a new Instance of a Web Context
 	 * @since 7.0.0
-	*/ constructor(controller: Server<any>, localContext: LocalContext, ws: WebSocket<WebSocketContext>) {
+	*/ constructor(controller: Server<any, any>, localContext: LocalContext, ws: WebSocket<WebSocketContext>) {
 		super(controller, localContext)
 
 		this.rawWs = ws